# openapi.yaml
openapi: 3.0.3
info:
  title: Cyber Storefront API
  version: "1.0.0"
  description: |
    **Cyber** ონლაინ მაღაზიის API: კატალოგი, ავტორიზაცია, კალათა, ჩექაუთი და პროფილი.

    ---

    # 🚀 სწრაფი დაწყება — სცადე აქვე, 4 ნაბიჯად

    ამ გვერდზე ყველა ენდპოინტის გატესტვა შეგიძლია, კოდის დაწერის გარეშე.

    ### ნაბიჯი 1 — შექმენი ანგარიში

    ჩამოდი `POST /auth/register`-მდე → დააჭირე **`Try it out`** (მარჯვნივ)
    → შეცვალე მონაცემები → **`Execute`**

    ```json
    { "name": "Test User", "email": "შენი@example.com", "password": "Password1" }
    ```

    ### ნაბიჯი 2 — დააკოპირე ტოკენი

    პასუხში (Response body) ნახავ:

    ```json
    {
      "user": { "id": "...", "name": "Test User", "email": "..." },
      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOi..."
    }
    ```

    დააკოპირე **`accessToken`-ის მნიშვნელობა** — ბრჭყალების გარეშე.

    ### ნაბიჯი 3 — დააჭირე `Authorize` 🔓

    ამ გვერდის **ზედა მარჯვენა კუთხეში** არის ღილაკი **`Authorize`** (ბოქლომის ხატულით).

    1. დააჭირე
    2. ველში `Value` ჩასვი დაკოპირებული ტოკენი
       *(მხოლოდ ტოკენი — სიტყვა `Bearer` ჩაწერა არ გჭირდება, Swagger თავად დაამატებს)*
    3. **`Authorize`** → **`Close`**

    ბოქლომები 🔒 დაიკეტება — ესე იგი ავტორიზებული ხარ.

    ### ნაბიჯი 4 — სცადე დაცული ენდპოინტი

    ჩამოდი `GET /auth/me`-მდე → **`Try it out`** → **`Execute`**

    ✅ დაბრუნდება შენი მომხმარებლის მონაცემები.

    **სცადე საწინააღმდეგოც:** დააჭირე `Authorize` → **`Logout`** → ისევ გაუშვი
...

    ---

...

...
servers:
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: რეგისტრაცია
      description: |
        👉 **პასუხიდან დააკოპირე `accessToken`** და ჩასვი ზემოთ ღილაკში `Authorize` 🔓 —
        ამის შემდეგ დაცული ენდპოინტების ტესტირებას შეძლებ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, minLength: 2, example: "Nino Beridze" }
                email: { type: string, format: email, example: "nino@example.com" }
                password:
                  type: string
                  minLength: 8
                  description: მინ. 8 სიმბოლო, მინიმუმ 1 ასო და 1 ციფრი
                  example: "Password1"
      responses:
        "201":
          description: ანგარიში შეიქმნა
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "409":
          description: ელფოსტა უკვე დაკავებულია (`EMAIL_TAKEN`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "422":
          $ref: "#/components/responses/Validation"
  /auth/login:
    post:
      tags: [Auth]
      summary: შესვლა
      description: |
        👉 **პასუხიდან დააკოპირე `accessToken`** და ჩასვი ზემოთ ღილაკში `Authorize` 🔓

        ⚠️ შესვლაზე პაროლის სირთულე **არ მოწმდება** — მხოლოდ სავალდებულოობა.
        მომხმარებლის არსებული პაროლი შეიძლება ძველი პოლიტიკით იყოს შექმნილი.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: "nino@example.com" }
                password: { type: string, example: "Password1" }
      responses:
        "200":
          description: შესვლა წარმატებულია
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: არასწორი ელფოსტა ან პაროლი (`INVALID_CREDENTIALS`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          $ref: "#/components/responses/Validation"
  /auth/me:
    get:
      tags: [Auth]
      summary: მიმდინარე მომხმარებელი
      description: |
        🔒 **საჭიროებს ტოკენს.** ჯერ გაიარე `Authorize` (იხ. სწრაფი დაწყება ზემოთ),
        თორემ მიიღებ `401 UNAUTHORIZED`.

        **რისთვის სჭირდება ფრონტენდს:** აპლიკაციის გაშვებისას ეს ერთი მოთხოვნა
        ერთდროულად პასუხობს ორ კითხვას — *„ჩემი ტოკენი ჯერ კიდევ ვარგა?"* და
        *„ვინ არის შესული მომხმარებელი?"*

        - `200` → ტოკენი ვალიდურია, აი მომხმარებელი
        - `401` → წაშალე ტოკენი `localStorage`-დან და გადაამისამართე `/login`-ზე
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: ტოკენი არ არის, არასწორია ან ვადაგასულია (`TOKEN_EXPIRED`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      tags: [Auth]
      summary: პროფილის რედაქტირება
      description: |
        🔒 **საჭიროებს ტოკენს.** გაგზავნე მხოლოდ ის ველები, რომლის შეცვლაც გინდა.

        **`currentPassword` ყოველთვის სავალდებულოა** — ნებისმიერ ცვლილებამდე სერვერი
        ამოწმებს მიმდინარე პაროლს. არასწორზე: `400 INVALID_CURRENT_PASSWORD` (და არა `401`,
        რომ ფრონტენდმა მომხმარებელი სისტემიდან არ გააგდოს).

        - `newPassword` — ახალი პაროლი (იგივე წესები, რაც რეგისტრაციაზე)
        - `phone`, `city`, `address` — `""` ან `null` ასუფთავებს ველს
        - ეს ველები ჩექაუთზე მისამართს ავტომატურად ავსებს (`GET /checkout`)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword]
              properties:
                currentPassword: { type: string, example: "Password1" }
                name: { type: string, minLength: 2, example: "Nino Beridze" }
                email: { type: string, format: email, example: "nino.new@example.com" }
                newPassword: { type: string, minLength: 8, example: "NewPass123" }
                phone: { type: string, nullable: true, example: "+995 555 12 34 56" }
                city: { type: string, nullable: true, example: "Tbilisi" }
                address: { type: string, nullable: true, example: "Rustaveli Ave 12, apt 5" }
      responses:
        "200":
          description: პროფილი განახლდა
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400":
          description: მიმდინარე პაროლი არასწორია (`INVALID_CURRENT_PASSWORD`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: ახალი ელფოსტა სხვა ანგარიშზეა (`EMAIL_TAKEN`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "422":
          $ref: "#/components/responses/Validation"
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: "აღდგენა — ნაბიჯი 1: კოდის მოთხოვნა"
      description: |
        👉 **პასუხიდან დააკოპირე `devCode`** — ის დაგჭირდება შემდეგ ნაბიჯზე
        (`/auth/verify-reset-code`).

        **ყოველთვის აბრუნებს 200-ს**, არსებობს ანგარიში თუ არა — თორემ ფორმა
        იქცევა ინსტრუმენტად, რომლითაც გაიგებ, ვის აქვს ანგარიში საიტზე.

        კოდი 6-ციფრიანია, ვადა 15 წუთი. ელფოსტა არ იგზავნება: სასწავლო რეჟიმში
        კოდი ბრუნდება `devCode`-ში.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, example: "nino@example.com" }
      responses:
        "200":
          description: ნეიტრალური პასუხი
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "If an account exists for this email, we have sent a 6-digit code." }
                  expiresInMinutes: { type: integer, example: 15 }
                  devCode: { type: string, example: "483920", description: "მხოლოდ სასწავლო რეჟიმში" }
        "422":
          $ref: "#/components/responses/Validation"
  /auth/verify-reset-code:
    post:
      tags: [Auth]
      summary: "აღდგენა — ნაბიჯი 2: კოდის შემოწმება"
      description: |
        👉 **პასუხიდან დააკოპირე `resetToken`** — ის დაგჭირდება ბოლო ნაბიჯზე
        (`/auth/reset-password`).

        ⚠️ ეს **არ არის** `accessToken`. `Authorize`-ში მისი ჩასმა არ გჭირდება —
        ის პირდაპირ სხეულში (body) მიდის.

        `resetToken`-ის ვადა 10 წუთია და **ერთჯერადია**.
        5 არასწორი მცდელობის შემდეგ კოდი იბლოკება (`429`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email, example: "nino@example.com" }
                code: { type: string, pattern: "^\\d{6}$", example: "483920" }
      responses:
        "200":
          description: კოდი სწორია
          content:
            application/json:
              schema:
                type: object
                properties:
                  resetToken: { type: string }
        "400":
          description: კოდი არასწორია ან ვადაგასულია (`INVALID_RESET_CODE`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "429":
          description: ძალიან ბევრი მცდელობა (`TOO_MANY_ATTEMPTS`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          $ref: "#/components/responses/Validation"
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: "აღდგენა — ნაბიჯი 3: ახალი პაროლი"
      description: |
        ჩასვი წინა ნაბიჯზე მიღებული `resetToken` და ახალი პაროლი.

        წარმატების შემდეგ სცადე `POST /auth/login` **ახალი** პაროლით —
        და დარწმუნდი, რომ ძველი აღარ მუშაობს.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resetToken, password]
              properties:
                resetToken: { type: string }
                password: { type: string, minLength: 8, example: "NewPassword1" }
      responses:
        "200":
          description: პაროლი შეიცვალა
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "400":
          description: ტოკენი არასწორია, ვადაგასულია ან უკვე გამოყენებულია
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          $ref: "#/components/responses/Validation"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: ტოკენი არ არის, არასწორია ან ვადაგასულია (`UNAUTHORIZED` / `TOKEN_EXPIRED`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Validation:
      description: ვალიდაციის შეცდომა (`VALIDATION_ERROR`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
  schemas:
    User:
      type: object
      properties:
        id: { type: string, example: "clx8f2k9a0000..." }
        name: { type: string, example: "Nino Beridze" }
        email: { type: string, example: "nino@example.com" }
        phone: { type: string, nullable: true, example: "+995 555 12 34 56" }
        city: { type: string, nullable: true, example: "Tbilisi" }
        address: { type: string, nullable: true, example: "Rustaveli Ave 12, apt 5" }
        createdAt: { type: string, format: date-time }
...
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        accessToken:
          type: string
          description: "ატარე ჰედერში: Authorization: Bearer <accessToken>"
    Error:
      type: object
      properties:
        message: { type: string, example: "Invalid email or password" }
        code: { type: string, example: "INVALID_CREDENTIALS" }
    ValidationError:
      type: object
      properties:
        message: { type: string, example: "Validation failed" }
        code: { type: string, example: "VALIDATION_ERROR" }
        errors:
          type: object
          additionalProperties: { type: string }
          example: { email: "Enter a valid email address", password: "Password must be at least 8 characters" }